Function loadImageFromUrl shifted to FrescoImageLoader class
