+++ app/src/main/java/fr/free/nrw/commons/widget/PicOfDayAppWidget.java
@@ -4,24 +4,11 @@ import android.appwidget.AppWidgetManager;
 import android.appwidget.AppWidgetProvider;
 import android.content.Context;
 import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.Paint;
-import android.net.Uri;
-import android.support.annotation.Nullable;
 import android.widget.RemoteViews;
-
-import com.facebook.common.executors.CallerThreadExecutor;
-import com.facebook.common.references.CloseableReference;
-import com.facebook.datasource.DataSource;
-import com.facebook.drawee.backends.pipeline.Fresco;
-import com.facebook.imagepipeline.core.ImagePipeline;
-import com.facebook.imagepipeline.datasource.BaseBitmapDataSubscriber;
-import com.facebook.imagepipeline.image.CloseableImage;
-import com.facebook.imagepipeline.request.ImageRequest;
-import com.facebook.imagepipeline.request.ImageRequestBuilder;
 
 import javax.inject.Inject;
 
+import fr.free.nrw.commons.FrescoImageLoader;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.di.ApplicationlessInjection;
 import fr.free.nrw.commons.mwapi.OkHttpJsonApiClient;
@@ -61,48 +48,13 @@ public class PicOfDayAppWidget extends AppWidgetProvider {
                 .subscribe(
                         response -> {
                             if (response != null) {
-                                loadImageFromUrl(response.getImageUrl(), context, views, appWidgetManager, appWidgetId);
+                                Bitmap bitmap = FrescoImageLoader.loadImageFromUrl(response.getImageUrl(), context);
+                                views.setImageViewBitmap(R.id.appwidget_image, bitmap);
+                                appWidgetManager.updateAppWidget(appWidgetId, views);
                             }
                         },
                         t -> Timber.e(t, "Fetching picture of the day failed")
                 ));
-    }
-
-    /**
-     * Uses Fresco to load an image from Url
-     * @param imageUrl
-     * @param context
-     * @param views
-     * @param appWidgetManager
-     * @param appWidgetId
-     */
-    private void loadImageFromUrl(String imageUrl,
-                                  Context context,
-                                  RemoteViews views,
-                                  AppWidgetManager appWidgetManager,
-                                  int appWidgetId) {
-        ImageRequest request = ImageRequestBuilder.newBuilderWithSource(Uri.parse(imageUrl)).build();
-        ImagePipeline imagePipeline = Fresco.getImagePipeline();
-        DataSource<CloseableReference<CloseableImage>> dataSource
-                = imagePipeline.fetchDecodedImage(request, context);
-        dataSource.subscribe(new BaseBitmapDataSubscriber() {
-            @Override
-            protected void onNewResultImpl(@Nullable Bitmap tempBitmap) {
-                Bitmap bitmap = null;
-                if (tempBitmap != null) {
-                    bitmap = Bitmap.createBitmap(tempBitmap.getWidth(), tempBitmap.getHeight(), Bitmap.Config.ARGB_8888);
-                    Canvas canvas = new Canvas(bitmap);
-                    canvas.drawBitmap(tempBitmap, 0f, 0f, new Paint());
-                }
-                views.setImageViewBitmap(R.id.appwidget_image, bitmap);
-                appWidgetManager.updateAppWidget(appWidgetId, views);
-            }
-
-            @Override
-            protected void onFailureImpl(DataSource<CloseableReference<CloseableImage>> dataSource) {
-                // Ignore failure for now.
-            }
-        }, CallerThreadExecutor.getInstance());
     }
 
     @Override
